initramfs-tools-tcos (0.89.4)   * xmlrpc/xauth.c xmlrpc/common.h     - Change mkstemp prototype, from man page: 'int mkstemp(char *template);'     - This fix amd64 warning of incompatible types.
git-svn-id: http://tcosproject.org/svn/tcos/trunk/initramfs-tools-tcos@923 97557d6f-af72-4ac7-98cc-18709f36555f

+++ xmlrpc/xauth.c
@@ -64,7 +64,7 @@ handle_xauth( char *cookie , char *servername)
         snprintf(host, BSIZE, "%s", ip.ipstr);
     }
     
-    fd = (int) mkstemp(xauth_file);
+    fd = mkstemp(xauth_file);
     if (fd < 0) {
         unlink(xauth_file);
         close(fd);
